WIP longest increasing subsequence

+++ include/push_swap.h
@@ -50,7 +50,7 @@ typedef struct s_push_swap {
 	t_item			*stack_a_index_head;
 	t_instruction	*instructions;
 	int				rotation_offset;
-	int				num_elements;
+	long			num_elements;
 }	t_push_swap;
 
 // An enum containing the different sorting operations available to us. Used
